refactor(subway/domain): package 분리
