+++ src/main/java/wooteco/subway/dao/LineDao.java
@@ -1,7 +1,7 @@
 package wooteco.subway.dao;
 
 import java.util.List;
-import wooteco.subway.domain.Fare;
+import wooteco.subway.domain.path.Fare;
 import wooteco.subway.domain.Line;
 
 public interface LineDao {
