+++ src/main/java/wooteco/subway/domain/Line.java
@@ -2,6 +2,8 @@ package wooteco.subway.domain;
 
 import java.util.LinkedList;
 import java.util.List;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.section.Sections;
 
 public class Line {
