+++ src/main/java/wooteco/subway/domain/section/Section.java
@@ -1,4 +1,6 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.section;
+
+import wooteco.subway.domain.Station;
 
 public class Section {
     private final Long id;
