+++ src/main/java/wooteco/subway/ui/service/PathService.java
@@ -7,14 +7,13 @@ import java.util.stream.Collectors;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.springframework.stereotype.Service;
 
-import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Fare;
-import wooteco.subway.domain.Path;
-import wooteco.subway.domain.PathEdge;
-import wooteco.subway.domain.Section;
-import wooteco.subway.domain.ShortestPathFactory;
+import wooteco.subway.domain.path.Fare;
+import wooteco.subway.domain.path.Path;
+import wooteco.subway.domain.path.PathEdge;
+import wooteco.subway.domain.section.Section;
+import wooteco.subway.domain.path.ShortestPathFactory;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.response.PathResponse;
 import wooteco.subway.dto.response.StationResponse;
