+++ src/main/java/wooteco/subway/ui/service/SectionService.java
@@ -5,9 +5,9 @@ import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.dao.LineDao;
 import wooteco.subway.dao.SectionDao;
 import wooteco.subway.dao.StationDao;
-import wooteco.subway.domain.Distance;
+import wooteco.subway.domain.section.Distance;
 import wooteco.subway.domain.Line;
-import wooteco.subway.domain.Section;
+import wooteco.subway.domain.section.Section;
 import wooteco.subway.domain.Station;
 import wooteco.subway.dto.request.SectionRequest;
