+++ src/test/java/wooteco/subway/domain/LineTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.domain.section.Distance;
+import wooteco.subway.domain.section.Section;
 
 class LineTest {
